🚚rename: 마이페이지 및 주식 페이지 폴더명 수정
