+++ frontend/src/routes/myPageF/HoldStock.jsx
