+++ frontend/src/routes/stockPageF/ChartPage.jsx
